Created the logic for registering new users.
The login in feature is almost to a point where it can be tested. Currently troubleshooting errors with the authentication middleware.
Error currently reads secretKey or PritaveKey must be provided. Current guess is an issue with the .env file.

Current userModel handles the schema for creating a new user, requiring a Unique username, unique email, and password. No error validation has been added to this section yet.
SQL sterilization will also need to be added at some point in this process.

Currently the .JS as it is will create a new table in our exisiting database under the name "users". This can be tailored to our needs once we have table names finalized. It also includes
createdAt and updatedAt columns.